[UBSan] Fix typo in CMake conditional that checked if the architecture of a darwin platform was in the list of `UBSAN_SUPPORTED_ARCH`.
This is a follow up to r341306.

The typo meant that if an architecture was a prefix to another
architecture in the list (e.g. `armv7` is a prefix of `armv7k`) then
this would trigger a match which is not the intended behaviour.

rdar://problem/41126835

git-svn-id: https://llvm.org/svn/llvm-project/compiler-rt/trunk@342553 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 01c0f7abd9218a7c83cda551cf6b861f3ac69407)