[LibFuzzer] Unbreak `fuzzer/fuzzer-oom.test`.
The test was compiling binaries with a `.exe` extension but then
tried running them without the extension. On incremental bots we
wouldn't notice the failure because we'd just run existing binaries.

It looks like this might be the result of a bad merge (possibly with
https://reviews.llvm.org/D51735) because
10b49e5094975a1a32ccd160383004ce30e9b4cd (the responsible commit)
doesn't look like https://reviews.llvm.org/D54004 as described in that
commit's description.

rdar://problem/46987680